align component code more closely with post-refactor element code
